刷新oss对象时判断 原对象是否存在

+++ app/Cdn/Adapter/AliOss.php
@@ -97,7 +97,9 @@ class AliOss implements FileUrlGeneratorContract
                 $delete[] = $object->getKey();
             }
 
-            $this->client->deleteObjects($this->bucket, $delete);
+            if ($delete) {
+                $this->client->deleteObjects($this->bucket, $delete);
+            }
         }
     }
